用信号触发排行榜更新
